move more Web APIs to cli/js/web/
