+++ cli/js/web/README.md
@@ -1,0 +1,8 @@
+# Deno Web APIs
+
+This directory facilities Web APIs that are available in Deno.
+
+Please note, that some of implementations might not be completely aligned with
+specification.
+
+Some of the Web APIs are using ops under the hood, eg. `console`, `performance`.
