+++ cli/js/web/timers.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
-import { assert } from "./util.ts";
-import { startGlobalTimer, stopGlobalTimer } from "./ops/timers.ts";
-import { RBTree } from "./rbtree.ts";
+import { assert } from "../util.ts";
+import { startGlobalTimer, stopGlobalTimer } from "../ops/timers.ts";
+import { RBTree } from "../rbtree.ts";
 
 const { console } = globalThis;
